Remove copyright headers, fully rename ztest to scalaz-test
