+++ scalaz-test/src/main/scala/ztest/PureSuite.scala
