+++ scalaz-test/src/main/scala/ztest/tests/PureSuiteMetaSuite.scala
