+++ scalaz-test/src/main/tut/docs/02-A-First-Example.md
