--- ztest/src/main/scala/ztest/data.scala
@@ -1,60 +1,0 @@
-/*
- * Copyright (c) 2018, Edmund Noble
- * All rights reserved.
- *
- * Redistribution and use in source and binary forms, with or without modification,
- * are permitted provided that the following conditions are met:
- *
- * 1. Redistributions of source code must retain the above copyright notice, this
- *    list of conditions and the following disclaimer.
- *
- * 2. Redistributions in binary form must reproduce the above copyright notice,
- *    this list of conditions and the following disclaimer in the documentation
- *    and/or other materials provided with the distribution.
- *
- * 3. Neither the name of the copyright holder nor the names of its contributors
- *    may be used to endorse or promote products derived from this software without
- *    specific prior written permission.
- *
- * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND
- * ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
- * WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
- * DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE LIABLE FOR
- * ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES
- * (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
- * LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON
- * ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
- * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
- * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
- */
-
-package scalaz.test
-
-import scala._, scala.Predef._
-
-import scalaz.Scalaz._, scalaz.core.EqClass
-
-trait Harness[F[_], G[_]] {
-  def section[A](annotation: String)(test: G[A]): G[A]
-  def shared[A](fa: F[A]): G[A]
-  def test(assertion: F[TestResult]): G[Unit]
-  final def namedTest(annotation: String)(errs: F[TestResult]): G[Unit] =
-    section(annotation)(test(errs))
-}
-
-sealed trait TestResult
-final class Thrown(val thrown: Throwable) extends TestResult
-object Thrown {
-  def apply(thrown: Throwable): TestResult = new Thrown(thrown)
-}
-object Success extends TestResult {
-  def apply(): TestResult = this
-}
-object Failure extends TestResult {
-  def apply(): TestResult = this
-}
-object TestResult {
-  implicit val equalTestResult: Eq[TestResult] = instanceOf(new EqClass[TestResult] {
-    def equal(first: TestResult, second: TestResult) = first == second
-  })
-}
