--- ztest/src/main/scala/ztest/property.scala
@@ -1,58 +1,0 @@
-/*
- * Copyright (c) 2018, Edmund Noble
- * All rights reserved.
- *
- * Redistribution and use in source and binary forms, with or without modification,
- * are permitted provided that the following conditions are met:
- *
- * 1. Redistributions of source code must retain the above copyright notice, this
- *    list of conditions and the following disclaimer.
- *
- * 2. Redistributions in binary form must reproduce the above copyright notice,
- *    this list of conditions and the following disclaimer in the documentation
- *    and/or other materials provided with the distribution.
- *
- * 3. Neither the name of the copyright holder nor the names of its contributors
- *    may be used to endorse or promote products derived from this software without
- *    specific prior written permission.
- *
- * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND
- * ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
- * WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
- * DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE LIABLE FOR
- * ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES
- * (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
- * LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON
- * ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
- * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
- * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
- */
-
-package scalaz.test
-
-import scalaz.{Applicative, Traversable}
-import scalaz.Scalaz._
-
-object property {
-
-  def exhaustive[F[_]: Applicative, T[_]: Traversable, I, O]
-                (in: T[(() => I)])
-                (testGenerator: I => F[O])
-                : F[T[O]] = {
-    Traversable[T].traverse(in)(i => testGenerator(i()))
-  }
-
-  def lazify[I](i: I): () => I = () => i
-
-  def exhaustiveS[F[_]: Applicative, T[_]: Traversable, I, O]
-                 (in: T[I])
-                 (testGenerator: I => F[O])
-                : F[T[O]] =
-    Traversable[T].traverse(in)(testGenerator)
-
-  // def exhaustiveV[F[_]: Applicative, I, O]
-  //                (in: (() => I)*)
-  //                (testGenerator: I => F[O]): F[O] =
-  //   exhaustive(in.toList)(testGenerator)
-
-}
