--- ztest/src/main/scala/ztest/tests/ExhaustiveSuite.scala
@@ -1,59 +1,0 @@
-/*
- * Copyright (c) 2018, Edmund Noble
- * All rights reserved.
- *
- * Redistribution and use in source and binary forms, with or without modification,
- * are permitted provided that the following conditions are met:
- *
- * 1. Redistributions of source code must retain the above copyright notice, this
- *    list of conditions and the following disclaimer.
- *
- * 2. Redistributions in binary form must reproduce the above copyright notice,
- *    this list of conditions and the following disclaimer in the documentation
- *    and/or other materials provided with the distribution.
- *
- * 3. Neither the name of the copyright holder nor the names of its contributors
- *    may be used to endorse or promote products derived from this software without
- *    specific prior written permission.
- *
- * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND
- * ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
- * WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
- * DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE LIABLE FOR
- * ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES
- * (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
- * LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON
- * ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
- * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
- * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
- */
-
-package scalaz.test
-package tests
-
-import scala._
-
-import scalaz.data.IList
-import scalaz.Monad
-import scalaz.Scalaz._
-import property._
-
-final class ExhaustiveSuite extends PureSuite {
-  def doTests[G[_]: Monad](harness: Harness[() => ?, G]): G[Unit] = {
-    import harness._
-    section("exhaustiveS int range")(
-      test {
-        val ls = IList(1, 2, 3, 4, 5, 6)
-        exhaustiveS(ls)(i =>
-          () => (i != 3)
-        ).map(errs =>
-          if (errs === IList(true, true, false, true, true, true)) {
-            Success()
-          } else {
-            Failure()
-          }
-        )
-      }
-    )
-  }
-}
